requirements.txt
alabaster==0.7.12
antlr4-python3-runtime==4.8
appnope==0.1.0
argon2-cffi==20.1.0
async-generator==1.10
attackcti==0.3.3
attrs==20.1.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
gitdb==4.0.5
GitPython==3.1.7
idna==2.10
imagesize==1.2.0
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-book==0.7.4
jupyter-cache==0.3.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyter-sphinx==0.2.4
latexcodec==2.0.1
libsass==0.20.0
markdown-it-py==0.4.9
MarkupSafe==1.1.1
mistune==0.8.4
myst-nb==0.8.5
myst-parser==0.9.1
nbclient==0.4.1
nbconvert==5.6.1
nbdime==2.0.0
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.3
oset==0.1.3
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.8.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pybtex==0.22.2
pybtex-docutils==0.2.2
pycparser==2.20
pydata-sphinx-theme==0.3.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
requests==2.24.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
smmap==3.0.4
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==2.4.4
sphinx-book-theme==0.0.35
sphinx-copybutton==0.3.0
sphinx-thebe==0.0.7
sphinx-togglebutton==0.2.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==1.0.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.19
stix2==2.0.2
stix2-patterns==1.3.1
taxii2-client==2.2.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
